bug(racingcar) : fix Car module folder position
