+++ src/App.js
@@ -1,5 +1,4 @@
-import { Console } from "@woowacourse/mission-utils";
-import Car from "./Car";
+import { Console, Random } from "@woowacourse/mission-utils";
 
 class App {
   constructor() {
@@ -104,6 +103,30 @@ class App {
     }
 
     return roundCount;
+  }
+}
+
+class Car {
+  constructor(name) {
+    this.name = name;
+    this.position = 0;
+  }
+
+  //UI 로직
+  getCarName() {
+    return this.name;
+  }
+
+  getCarPosition() {
+    return this.position;
+  }
+
+  //비지니스 로직
+  checkPosition() {
+    const randomString = Random.pickNumberInRange(0, 9);
+    if (randomString >= 4) {
+      this.position += 1;
+    }
   }
 }
